DRY this up a bit
Call visitLandingPageAndCompleteCheckout directly from tests.
The intermediate methods weren't really adding anything any more.
