+++ support-e2e/tests/smoke/newspaper.test.ts
@@ -1,6 +1,14 @@
 import test from '@playwright/test';
-import { testNewspaperCheckout } from '../test/newspaperCheckout';
-import type { TestDetails } from '../test/newspaperCheckout';
+import { visitLandingPageAndCompleteCheckout } from '../utils/visitLandingPageAndCompleteCheckout';
+
+type TestDetails = {
+	product: string;
+	productLabel: string;
+	ratePlanLabel: string;
+	paymentType: string;
+	internationalisationId: string;
+	postCode?: string;
+};
 
 const tests: TestDetails[] = [
 	{
@@ -29,4 +37,41 @@ const tests: TestDetails[] = [
 ];
 
 test.describe('Newspaper Checkout', () =>
-	tests.map((testDetails) => testNewspaperCheckout(testDetails)));
+	tests.map((testDetails: TestDetails) => {
+		const {
+			product,
+			productLabel,
+			paymentType,
+			internationalisationId,
+			postCode,
+			ratePlanLabel,
+		} = testDetails;
+
+		test(`Newspaper - ${product} - ${paymentType} - ${internationalisationId} ${postCode ? `- ${postCode}` : ''}`, async ({
+			context,
+			baseURL,
+		}) => {
+			await visitLandingPageAndCompleteCheckout(
+				`/${internationalisationId.toLowerCase()}/subscribe/paper`,
+				{
+					context,
+					baseURL,
+					product,
+					paymentType,
+					internationalisationId,
+					postCode,
+				},
+				async (page) => {
+					// Transition from landing page to checkout:
+
+					// 1. Select the product (Home Delivery or Subscription Card)
+					await page.getByRole('tab', { name: productLabel }).click();
+
+					// 2. Click through to the checkout (we use the aria-label to target the link)
+					await page
+						.getByLabel(`${ratePlanLabel}- Subscribe`, { exact: true })
+						.click();
+				},
+			);
+		});
+	}));
