+++ support-e2e/tests/smoke/three-tier.test.ts
@@ -1,6 +1,13 @@
 import test from '@playwright/test';
-import { testThreeTierCheckout } from '../test/threeTierCheckout';
-import type { TestDetails } from '../test/threeTierCheckout';
+import { visitLandingPageAndCompleteCheckout } from '../utils/visitLandingPageAndCompleteCheckout';
+
+export type TestDetails = {
+	productLabel: string;
+	paymentType: string;
+	product: string;
+	billingFrequency: string;
+	internationalisationId: string;
+};
 
 const tests: TestDetails[] = [
 	{
@@ -20,4 +27,31 @@ const tests: TestDetails[] = [
 ];
 
 test.describe('Three Tier Checkout', () =>
-	tests.map((testDetails) => testThreeTierCheckout(testDetails)));
+	tests.map((testDetails: TestDetails) => {
+		const {
+			billingFrequency,
+			product,
+			paymentType,
+			internationalisationId,
+			productLabel,
+		} = testDetails;
+
+		test(`Three Tier - ${product} - ${billingFrequency} - ${paymentType} - ${internationalisationId}`, async ({
+			context,
+			baseURL,
+		}) => {
+			await visitLandingPageAndCompleteCheckout(
+				`/${internationalisationId.toLowerCase()}/contribute`,
+				{ context, baseURL, product, paymentType, internationalisationId },
+				async (page) => {
+					// Transition from landing page to checkout:
+
+					// 1. Select the billing frequency
+					await page.getByRole('tab', { name: billingFrequency }).click();
+
+					// 2. Click through to the checkout (we use the aria-label to target the link)
+					await page.getByLabel(productLabel, { exact: true }).click();
+				},
+			);
+		});
+	}));
